Reworked _find_required_grid_attrs to return generated topology
If a topology is provided or discovered in the file, this overrides any
generated topology (even if it may be incomplete).

When a Dataset is created from a file, any variables that were used to
create the associated grid are ignored when creating it's Variable list.
This behavior is determined by the Dataset.grid.grid_topology. Any
netCDF.Variable names appearing in it are IGNORED.

From there, of the remaning netCDF.Variable names, only ones that are
correctly shaped to fit the grid are accepted.